in progress refactor
